separates basic and business data into own find handlers
