+++ cg/store/api/core.py
@@ -4,17 +4,18 @@ import logging
 import alchy
 
 from cg.store import models
-from cg.store.api.find2 import FindHandler2
+from cg.store.api.findbusinessdata import FindBusinessDataHandler
 
 from .add import AddHandler
-from .find import FindHandler
+from .findbasicdata import FindBasicDataHandler
 from .status import StatusHandler
 from .trends import TrendsHandler
 
 LOG = logging.getLogger(__name__)
 
 
-class CoreHandler(AddHandler, FindHandler, FindHandler2, StatusHandler, TrendsHandler):
+class CoreHandler(AddHandler, FindBasicDataHandler, FindBusinessDataHandler, StatusHandler,
+                  TrendsHandler):
     """Aggregating class for the store api handlers"""
     pass
